+++ src/Upgrades/Upgrade27.php
@@ -13,10 +13,10 @@ class Upgrade27 implements IUpgrade
 			'CREATE TABLE postNotes (
 				id INTEGER PRIMARY KEY ' . ($driver === 'mysql' ? 'AUTO_INCREMENT' : 'AUTOINCREMENT') . ',
 				postId INTEGER NOT NULL,
-				x INTEGER NOT NULL,
-				y INTEGER NOT NULL,
-				height INTEGER NOT NULL,
-				width INTEGER NOT NULL,
+				x DECIMAL(6,2) NOT NULL,
+				y DECIMAL(6,2) NOT NULL,
+				height DECIMAL(6,2) NOT NULL,
+				width DECIMAL(6,2) NOT NULL,
 				text TEXT NOT NULL
 			)');
 	}
